Add Launch Debug Adapter menu
The menu does not do much. It allows a launch.json file to be
selected. The file selected is read and converted into a class.
The launch.json file format supported is not the same as the one
that VS Code uses. The format currently only allows one launch
configuration to be specified:

{
  "$adapter": "bin/Debug/SampleDebugAdapter.exe",

  "type": "sample",
  "request": "launch",
  "program": "src/sample/SampleDebugAdapter/SampleScripts/TestScript.txt"
}