Debug breakpoints now working
When the initialize event is received the DebuggerSession's OnStarted
method is now called. The OnStarted method will enable the breakpoints.
Unfortunately the VSCodeDebuggerSession does not wait until the
breakpoints have been configured so added a hack where the code waits
for a second before sending the configuration done request otherwise
the debugger runs and misses the breakpoints.

Also fixed an IndexOutOfRange exception when accesses the process
information. Some debug adapters, such as node, do not provide the
process information so the code now returns a dummy process.