Support debugger tooltips
Hovering an item will now show a debug value tooltip if the debugger
is currently paused.

This is a bit of a hack. The DebugValueTooltipProvider only works
if the DocumentContext.ParsedDocument is not null. If it is null
then the IDebuggerExpressionResolver that is implemented on the
text editor extension is not used. As a workaround a custom tooltip
provider is defined which finds the existing DebugValueTooltipProvider
from the extensions, then adds a ParsedDocument to the DocumentContext
if it is missing, and then forwards the method calls onto the
DebugValueTooltipProvider.